Major refactoring on Kosa project.
