+++ about.php
@@ -1,129 +1,24 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+<!-- ----------------------start of header.php  ------------------------------>
+<?php
 
-    <!--bootstrap  css-->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-BmbxuPwQa2lc/FVzBcNJ7UAyJxM6wuqIj61tLrc4wSX0szH/Ev+nYRRuWlolflfl" crossorigin="anonymous">
-    
-    <!-- local stylesheet-->
-    <link href="stylesheet/mainstyle.css" rel="stylesheet">
+  include('header.php');
 
-    <style>
+?>
+<!-- -----------------------end of header.php  ------------------------------->
 
-.top_container{
-  background:url(./images/adult.jpg);
-  background-size:cover;
-    background-position: center;
-    height:100vh;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-  
-}
-.top_container h1,p{
-  color:#fff;
-  background:#0004;
-  font-family:cambria;
-}
-.top_container h1{
- font-size:58px;
+   <!-- -----------------------start of about-us template  --------------------------->
+ <?php
+
+     include('templates/_about-us.php');
+
+?>
+<!-- -----------------------------------end of about-us template  -------------------->
  
-}
-.top_container p{
- font-size:24px;
-  
-}
+   
+   <!-- -----------------------start of footer.php  --------------------------->
+<?php
 
+include('footer.php');
 
-
-    </style>
-
-    <title>Home|About</title>
-</head>
-<body>
-                      <!-- start of navbar  -->
-  <nav class="navbar navbar-expand-lg ">
-    <div class="container-fluid nav_div">
-      <a class="navbar-brand" href="index.php">KOSA</a>
-      <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-      </button>
-      <div class="collapse navbar-collapse knav" id="navbarSupportedContent">
-        <ul class="navbar-nav me-auto mb-2 mb-lg-0">
-          <li class="nav-item">
-            <a class="nav-link active" aria-current="page" href="index.php">Home</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="#">Forum</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="about.php">About Us</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="alumnilogin.php">Login</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="register.php">Register</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="#">Contact Admins</a>
-          </li>
-
-          <!-- <li class="nav-item dropdown">
-            <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-              Dropdown
-            </a>
-            <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
-              <li><a class="dropdown-item" href="#">Action</a></li>
-              <li><a class="dropdown-item" href="#">Another action</a></li>
-              <li><hr class="dropdown-divider"></li>
-              <li><a class="dropdown-item" href="#">Something else here</a></li>
-            </ul>
-          </li> -->
-        </ul>
-        <form class="d-flex">
-          <input class="form-control me-2" type="search" placeholder="Search" aria-label="Search">
-          <button class="btn search_btn" type="submit">Search</button>
-        </form>
-      </div>
-    </div>
-  </nav>
-                         <!-- end of navbar  -->
-
-                         <!-- start of top container  -->
-                         <div class="container-fluid top_container">
-                           <h1>WHO ARE WE .....?</h1>
-                           <p>KOSA(2008) is an Alumni group of Kasipe Memorial International School.
-                              As a group, we hold on to good old memories whilst we hope and pray for a brighter tomorrow. 
-                              We all strive to be there for each other in good or bad times. 
-                              After going through one of the biggest challenges of our life (Battle of ZAMA),
-                             we are always proud to call ourselves the DISCIPLES. 
-
-                           </p>
-
-
-                         </div>
-                         <!-- end of top container  -->
-
-
-
-
-                  <!-- start of footer  -->
-  <footer> &copy;KOSA <?php echo date("Y");?></footer>
-                  <!-- end of footer  -->
-
-
-
-
-
-  <!-- JavaScript Bundle with Popper -->
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/js/bootstrap.bundle.min.js" integrity="sha384-b5kHyXgcpbZJO/tY9Ul7kGkf1S0CWuKcCD38l8YkeH8z8QjE0GmW1gYU5S9FOnJ0" crossorigin="anonymous"></script>
-
-</body>
-</html>
-
-
+?>
+<!-- -----------------------------------end of footer.php  -------------------->
