+++ admin/footer.php
@@ -1,0 +1,52 @@
+</main>
+
+<!--  ------------------------------------end of main  ---------------------------------------------->
+
+                   <!-- start of navbar  -->
+
+<!-- start of footer section  -->
+
+<section>
+
+
+               <!-- start of footer  -->
+               <footer> &copy;KOSA <?php echo date("Y");?></footer>
+               <!-- end of footer  -->
+
+
+
+ <!-- ****************************  end of default hidden contents****************************  -->
+
+               <!-- start of form control script -->
+               <script>
+
+<!-- hide registration form-->
+  function hide_form()
+  {
+    
+    var registration_form = document.getElementById("regform");
+    registration_form.style.display="none";
+  
+  }
+  function show_form()
+  {
+    var registration_form = document.getElementById("regform");
+    registration_form.style.display="block";
+
+  }
+</script>
+      <!-- end of form control script -->
+
+<!-- JavaScript Bundle with Popper -->
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/js/bootstrap.bundle.min.js" integrity="sha384-b5kHyXgcpbZJO/tY9Ul7kGkf1S0CWuKcCD38l8YkeH8z8QjE0GmW1gYU5S9FOnJ0" crossorigin="anonymous"></script>
+<!-- jquery cdn  -->
+<script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.6.0/jquery.js" integrity="sha512-n/4gHW3atM3QqRcbCn6ewmpxcLAHGaDjpEBu4xZd47N0W2oQ+6q7oc3PXstrJYXcbNU1OHdQ1T7pAP+gi5Yu8g==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
+<!-- owl carousel cdn  -->
+<script src="https://cdnjs.cloudflare.com/ajax/libs/OwlCarousel2/2.3.4/owl.carousel.min.js" integrity="sha512-bPs7Ae6pVvhOSiIcyUClR7/q2OAsRiovw4vAkX+zJbw3ShAeeqezq50RIIcIURq7Oa20rW2n2q+fyXBNcU9lrw==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
+
+</section>
+<!-- end  of footer section  -->
+</body>
+</html>
+
+
