+++ functions.php
@@ -1,0 +1,14 @@
+<?php
+//database class file
+require('database/DBController.php');
+
+//DBController object
+$db = new DBController();
+
+
+
+
+
+
+
+?>
