+++ package.json
@@ -1,0 +1,20 @@
+{
+  "name": "kosa",
+  "version": "1.0.0",
+  "description": "An alumni website for old students of Kasipe Memorial International School.",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "sass --watch ./sass/style.scss style.css"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/kpodjison/Kosa-Alumni-website-.git"
+  },
+  "author": "Kpodji Kwasi Emmanuel (jeevista)",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/kpodjison/Kosa-Alumni-website-/issues"
+  },
+  "homepage": "https://github.com/kpodjison/Kosa-Alumni-website-#readme"
+}
