+++ sass/include/colors.scss
@@ -1,0 +1,13 @@
+// color template 
+.color-primary{
+    color:$color-kosa-primary;
+}
+.color-secondary{
+    color:$color-kosa-blue;
+}
+.bg-trasp{
+    background: $color-transp;
+}
+.bg-kosa-blue{
+    background: $color-kosa-blue;
+}
