+++ sass/include/footer.scss
@@ -1,0 +1,8 @@
+footer {
+    background: #000;
+    color: #00b8e6;
+    font-size: 28px;
+    height: 40px;
+    text-align: center;
+  }
+
