+++ sass/include/mainstyle.scss
@@ -1,0 +1,31 @@
+
+nav{
+    // background:#00b8e6;
+     padding: 0px;
+    a{
+        color:#000;
+        box-sizing: border-box;
+        font-size: 25px;
+
+    }
+}
+.nav-items{
+    padding: 30px;
+}
+body{
+    background-color: yellow;
+}
+.knav a:hover{
+    background:#000;
+   color: #00b8e6;
+   
+}
+.search_btn{
+    background: #000;
+    color: #00b8e6;
+}
+.search_btn:hover{
+    background:#fff;
+    color: #000;
+}
+
