+++ sass/include/typography.scss
@@ -1,0 +1,13 @@
+/*typograph template */
+.font-rale{
+    font-family: $font-raleway;
+}
+.font-rubik{
+    font-family: $font-rubik;
+}
+.font-baloo{
+    font-family: $font-baloo;
+}
+.font-acme{
+    font-family: $font-acme;
+}
