+++ sass/include/variables.scss
@@ -1,0 +1,11 @@
+/*font variable*/
+$font-acme: "Acme",cursive;
+$font-baloo:"Baloo Thambi 2",cursive;
+$font-raleway: "Raleway",cursive;
+$font-rubik: "Rubik",cursive;
+
+/*color variables*/
+$color-kosa-primary: #ffffff;
+$color-transp: #0004;
+$color-kosa-blue: #00b8e6;
+$admin-panel:#9d0004;
