+++ templates/_about-us.php
@@ -1,0 +1,19 @@
+
+
+                         <!-- start of top container  -->
+ <div class="container-fluid top_container">
+    <h1>WHO ARE WE .....?</h1>
+        <p>KOSA(2008) is an Alumni group of Kasipe Memorial International School.
+                              As a group, we hold on to good old memories whilst we hope and pray for a brighter tomorrow. 
+                              We all strive to be there for each other in good or bad times. 
+                              After going through one of the biggest challenges of our life (Battle of ZAMA),
+                             we are always proud to call ourselves the DISCIPLES. 
+
+         </p>
+
+
+</div>
+                         <!-- end of top container  -->
+
+
+
