+++ templates/_register.php
@@ -1,0 +1,42 @@
+<div class="wrapper">
+
+
+<div class="greetings"><p>WELCOME TO KOSA 2008( The Disciples ) LOGIN PAGE</p></div>
+      
+          <!-- start of registration form -->                          
+                      <div class="container innercontainer" >
+                                <p>Register For An Alumni Account</p>
+                                <div class="container formdiv" >
+
+
+                                    <form action="POST">
+                                        
+                                        <lable for="firstname">Firstname</lable>
+                                            <input type="text" class="form-control"  style="width:50%" name="firstname" placeholder="Enter firstname" required><br>
+                                        <lable >Lastname</lable>
+                                            <input type="text" class="form-control" style="width:50%"  name="lastname" placeholder="Enter Lastname" required><br>
+                                        <lable>PhoneNumber</lable>
+                                            <input type="text" class="form-control" style="width:50%"  name="phonenum" placeholder="Enter phonenumber" required><br>
+                                        <lable for="email">Email</lable>
+                                            <input type="email" class="form-control" style="width:50%"  name="email" placeholder="Enter your email" required><br>
+                                        <lable for="reg_code">Registration code</lable>
+                                            <input type="text" class="form-control" style="width:50%"  name="reg_code" placeholder="Enter registration code" required><br>
+                                        <lable for="password" placeholder="Enter password" >Password</lable>
+                                            <input type="password" class="form-control" style="width:50%"  name="password" placeholder="Enter password" required><br>
+                                        <lable>Comfirm Password</lable>
+                                            <input type="text" class="form-control" style="width:50%"  name="comfirm_password" placeholder="Comfirm password" required><br>                   
+                                            <input type="submit" class="form-control" value="REGISTER" style="background:#00b8e6;color:#000;padding:5px;font-size:25px;
+                                        border:none;border-radius:20px;width:150px">
+
+
+                                    </form>
+                                    <br>
+
+                                    <span>Already have an account?<a href="alumnilogin.php" >Login Now!</a></span> 
+                                
+                                </div>       
+    
+                      </div>
+        <!-- end of registration form -->  
+
+</div>
